Rename operation report files to due case
