+++ app/pages/due_cases_comp/__init__.py
@@ -1,3 +1,3 @@
-from .operational_report_body import operational_report_body
+from .due_case_body import due_case_body
 from .day_selector_row import day_selector_row
 from .counting_section import counting_section
