+++ app/pages/due_cases_comp/counting_section.py
